# docs/api/api.yaml
openapi: 3.0.0
info:
  version: '1'
  title: Social API
  description: >
    The Social API is a comprehensive RESTful interface developed and

    maintained by the Tranvelocity team. It facilitates communication with the

    Social platform, allowing users to access and manage various

    social-related functionalities.


    Authentication and Authorization:


    The API employs a robust authentication and authorization mechanism to

    ensure secure access to resources. To interact with the API, users must

    include the following headers in their requests:


    - `X-Social-Api-Key`: A unique API key assigned to each user for

    authentication purposes.


    - `X-Social-Timestamp`: A Unix timestamp indicating the time of the

    request.


    - `X-Social-Signature`: A cryptographic signature generated using the

    `SHA-512` algorithm based on the API key, timestamp, and secret key. These

    headers are required for all requests.


    - `X-Social-User-SSID`: The User login session ID for user

    authentication. If provided, the API will specify the user's role based on

    this login session.


    - `X-Social-Auth-Id`: The authentication ID providing additional user

    identification.


    The Social API will specify the user's role based on this login session.
    If these `X-Social-User-SSID` and `X-Social-Auth-Id`

    headers are not included, it indicates that the request is sent by

    non-registered users.


    Role-Based Access Control:


...
servers:
...
paths:
...
  '/posts/{postId}/comments':
    get:
      summary: Retrieve comments for a specific post
      description: >
        - This endpoint retrieves comments associated with a specific post.

        - Everyone do have permission to view comments.

        - Returns the latest 30 comments. To retrieve more comments, provide the
        `last_id` parameter instead of offset to paginate.

        - Please pass the ID of the oldest comment currently displayed via
        `last_id`,

        and fetch the next 30 comments after that ID.

        - Nickname and avatar information of the commenter are fetched from the
        CRM API and are not managed by the Social API.
      tags:
        - Comment-related endpoints
      operationId: retrieve-comments-for-specific-post
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: postId
          in: path
          description: The ID of the post to retrieve comments for
          required: true
          schema:
            type: string
            example: 01hnw7nqm19g9302c7wq391nvm
        - name: comment
          in: query
          description: Filter comments by content
          schema:
            type: string
        - name: is_hidden
          in: query
          description: Filter comments by visibility status
          schema:
            type: boolean
        - name: user_id
          in: query
          description: Filter comments by User ID
          schema:
            type: integer
        - name: last_id
          in: query
          description: Filter comments by last ID
          schema:
            type: string
      responses:
        '200':
          description: A list of comments associated with the specified post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentList'
              examples:
                comment_list:
                  value:
                    data:
                      - id: 01hqt9drrjar3rc0ge0zpxwv93
                        user_id: 16614
                        nickname: hoctran
                        avatar: >-
                          https://stage.user.com/file/fileconsumer/doo/selector/downloader/phase/download_file/type/images/file_id/phpv2w4cF
                        post_id: 01hqt819gw8q83hfedkpbq5cp0
                        comment: Your updated comment haha
                        is_hidden: false
                        updated_at: '2024-02-29 20:20:37'
                        created_at: '2024-02-29 20:19:05'
                      - id: 01hqt99m80ymwtk3gqhr42j2by
                        user_id: 16614
                        nickname: hoctran
                        avatar: >-
                          https://stage.user.com/file/fileconsumer/doo/selector/downloader/phase/download_file/type/images/file_id/phpv2w4cF
                        post_id: 01hqt819gw8q83hfedkpbq5cp0
                        comment: Yabai Japanese
                        is_hidden: false
                        updated_at: '2024-02-29 20:16:50'
                        created_at: '2024-02-29 20:16:50'
                    success: true
                    code: 200
                    pagination:
                      offset: 0
                      limit: 30
                      total: 2
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '404':
          $ref: '#/components/responses/404PostResourceNotFoundResponse'
    post:
      summary: Store a newly created comment
      description: >
        This endpoint allows users to create a new comment for a specific post.
        The Social API will deny the request if the `comment` parameter
        contains any prohibited words that were registered in the CRM API.

        Please note the following:

        - Only registered users with appropriate permissions (posters and paid
        members) can leave comments on posts.

        - Nickname and avatar information of the commenter are fetched from the
        CRM API and are not managed by the Social API.

        - The prohibited (NG) words are managed by the CRM API, not by the
        Social API. For more information on managing prohibited words, please
        refer to the [CRM API
        documentation](https://apidoc.hoctran.net/tranvelocity/crm-laravel/v1/staging/#tag/Group-NgWord/operation/post-1-batch-ng-words).

        - Only posters may leave comments containing NG words.

        - Restricted users cannot leave comments.

        - Users who exceed the allowed number of comments in a specific time
        frame (e.g., maximum of 2 comments per hour) must wait before commenting
        again.
      tags:
        - Comment-related endpoints
      operationId: create-new-comment
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: postId
          in: path
          description: The ID of the post for which the comment is created
          required: true
          schema:
            type: string
          example: 01hnw7nqm19g9302c7wq391nvm
      requestBody:
        $ref: '#/components/requestBodies/CreateCommentRequestBody'
      responses:
        '201':
          $ref: '#/components/responses/201CreatedCommentResponse'
        '400':
          description: >-
            Bad request - Missing required parameters or contains prohibited
            word
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                prohibited_word:
                  summary: Comment contains prohibited word
                  value:
                    success: false
                    code: 400
                    errors:
                      message: 'The comment contains a prohibited word: `バカ`.'
                      code: 400001
                validation_error:
                  summary: Validation error - Missing required parameter
                  value:
                    data: []
                    success: false
                    code: 400
                    errors:
                      message: required|The comment field is required.
                      validation:
                        - attribute: comment
                          errors:
                            - key: Required
                              message: required|The comment field is required.
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          description: Forbidden - Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates whether the request was successful
                    example: false
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 403
                  errors:
                    type: object
                    properties:
                      message:
                        type: string
                        description: >-
                          A message explaining the reason for the forbidden
                          access
                        example: コメントを新しく作成する権限がありません。
                      code:
                        type: integer
                        description: A code indicating the specific error
                        example: 403001
              examples:
                permission_denied:
                  summary: Permission denied to create a comment
                  value:
                    success: false
                    code: 403
                    errors:
                      message: コメントを新しく作成する権限がありません。
                      code: 403001
                restricted_user:
                  summary: User is restricted and cannot comment
                  value:
                    success: false
                    code: 403
                    errors:
                      message: あなたは制限ユーザーのため、コメントを投稿できません。
                      code: 403001
                rate_limit_exceeded:
                  summary: Comment rate limit exceeded
                  value:
                    success: false
                    code: 403
                    errors:
                      message: 指定された時間内に許可されたコメントの最大数に達しました。（1時間につき最大2回までコメントできます。）
                      code: 403001
        '404':
          $ref: '#/components/responses/404PostResourceNotFoundResponse'
  '/posts/{postId}/comments/{id}':
    get:
      summary: Retrieve a specific comment
      description: >
        - Retrieve the specified comment resource by ID.

        - Everyone can view published comments

        - Nickname and avatar information of the commenter are fetched from the
        CRM API and are not managed by the Social API.
      tags:
        - Comment-related endpoints
      operationId: retrieve-a-specific-comment-for-specific-post
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: postId
          in: path
          description: The ID of the post containing the comment.
          required: true
          schema:
            type: string
            example: 01hnw7nqm19g9302c7wq391nvm
        - name: id
          in: path
          description: The ID of the comment to retrieve.
          required: true
          schema:
            type: string
            example: 01hnwcp6kd9a50w0hy2q0tg0zd
      responses:
        '200':
          description: Comment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResource'
              example:
                data:
                  id: 01hpjnnfennwxzgreg2r5wtx7f
                  user_id: 16614
                  nickname: hoctran
                  avatar: >-
                    https://stage.user.com/file/fileconsumer/doo/selector/downloader/phase/download_file/type/images/file_id/phpv2w4cF
                  comment: Your comment goes here
                  is_hidden: false
                  updated_at: '2024-02-14 11:03:23'
                  created_at: '2024-02-14 11:03:23'
                  total_comments: 2
                  post:
                    id: 01hp47vb1hcr1bcdr3c79bx6xm
                    admin_uuid: 9295de24-404b-408f-8ec5-1a5f3224a244
                    type: premium
                    content: Test content
                    is_published: true
                    published_start_at: '2023-11-30 08:00:00'
                    published_end_at: '2024-12-01 18:00:00'
                    updated_at: '2024-02-08 20:34:12'
                    created_at: '2024-02-08 20:32:33'
                success: true
                code: 200
                pagination: null
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '404':
          description: Comment or post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                example-1:
                  value:
                    success: false
                    code: 404
                    errors:
                      message: ID 01hnw7nqm19g9302c7wq391nvm11に一致する投稿は見つかりませんでした。
                      code: 404001
                example-2:
                  value:
                    success: false
                    code: 404
                    errors:
                      message: ID 01hnw7nqm19g9302c7wq391nvm11に一致するコメントは見つかりませんでした。
                      code: 404001
    put:
      summary: Update a comment
      description: >
        - Update the specified comment in storage. - The Social API will deny
        the request if the `comment` parameter contains any prohibited words
        that were registered in the CRM API.

        - Posters have permission to update user's comments

        - Commenter who left the comment have permission to update their
        comments. A user cannot update comments of other commenters

        - Nickname and avatar information of the commenter are fetched from the
        CRM API and are not managed by the Social API.

        - The prohibited (NG) words are managed by the CRM API, not by the
        Social API. For more information on managing prohibited words, please
        refer to the [CRM API
        documentation](https://apidoc.hoctran.net/tranvelocity/crm-laravel/v1/staging/#tag/Group-NgWord/operation/post-1-batch-ng-words).

        - Only posters may leave comments containing NG words.
      tags:
        - Comment-related endpoints
      operationId: update-a-specific-comment
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: postId
          in: path
          description: The ID of the post associated with the comment.
          required: true
          schema:
            type: string
            example: 01hnw7nqm19g9302c7wq391nvm
        - name: id
          in: path
          description: The ID of the comment to update.
          required: true
          schema:
            type: string
            example: 01hnwcp6kd9a50w0hy2q0tg0zd
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResource'
              example:
                data:
                  id: 01hqt9drrjar3rc0ge0zpxwv93
                  user_id: 16614
                  nickname: hoctran
                  avatar: >-
                    https://stage.user.com/file/fileconsumer/doo/selector/downloader/phase/download_file/type/images/file_id/phpv2w4cF
                  comment: Your updated comment haha
                  is_hidden: false
                  updated_at: '2024-02-29 20:20:37'
                  created_at: '2024-02-29 20:19:05'
                  total_comments: 2
                  post:
                    id: 01hqt819gw8q83hfedkpbq5cp0
                    admin_uuid: 6c3a77e6-06cb-43f5-bb49-3b44ffc0eaaa
                    type: premium
                    content: Test content
                    is_published: true
                    published_start_at: '2023-11-30 08:00:00'
                    published_end_at: '2024-12-01 18:00:00'
                    updated_at: '2024-02-29 20:02:21'
                    created_at: '2024-02-29 19:54:48'
                success: true
                code: 200
        '400':
          description: >-
            Bad request - Missing required parameters or contains prohibited
            word
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                prohibited_word:
                  summary: Comment contains prohibited word
                  value:
                    success: false
                    code: 400
                    errors:
                      message: 'The comment contains a prohibited word: `バカ`.'
                      code: 400001
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '404':
          $ref: '#/components/responses/404CommentAndPostResourceNotFoundResponse'
    delete:
      summary: Delete a comment
      description: >
        Remove the specified comment from storage. Only administrators, posters,
        and commenters have permission to delete a comment.
      tags:
        - Comment-related endpoints
      operationId: delete-a-specific-comment
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: postId
          in: path
          description: The ID of the post associated with the comment.
          required: true
          schema:
            type: string
            example: 01hnw7nqm19g9302c7wq391nvm
        - name: id
          in: path
          description: The ID of the comment to delete.
          required: true
          schema:
            type: string
            example: 01hnwcp6kd9a50w0hy2q0tg0zd
      responses:
        '204':
          description: The media was successfully deleted
          content:
            application/json:
              schema:
                type: string
                example: No content
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '404':
          $ref: '#/components/responses/404CommentAndPostResourceNotFoundResponse'
  '/posts/{postId}/comments/{id}/publish':
    post:
      summary: Publish a comment
      description: >
        - Publishes a comment associated with a post.

        - Only posters have permission to publish a comment

        - Nickname and avatar information of the commenter are fetched from the
        CRM API and are not managed by the Social API.
      tags:
        - Comment-related endpoints
      operationId: publish-a-comment
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: postId
          in: path
          description: The ID of the post associated with the comment.
          required: true
          schema:
            type: string
            example: 01hnw7nqm19g9302c7wq391nvm
        - name: id
          in: path
          description: The ID of the comment to publish.
          required: true
          schema:
            type: string
            example: 01hnwcp6kd9a50w0hy2q0tg0zd
      responses:
        '200':
          description: Comment published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResource'
              example:
                data:
                  id: 01hnwcp6kd9a50w0hy2q0tg0zd
                  user_id: 16614
                  nickname: hoctran
                  avatar: >-
                    https://stage.user.com/file/fileconsumer/doo/selector/downloader/phase/download_file/type/images/file_id/phpv2w4cF
                  comment: Your comment goes here
                  is_hidden: false
                  updated_at: '2024-02-07 19:08:15'
                  created_at: '2024-02-05 19:23:13'
                  post:
                    id: 01hnw7nqm19g9302c7wq391nvm
                    admin_uuid: 9295de24-404b-408f-8ec5-1a5f3224a244
                    type: premium
                    content: Test content
                    is_published: true
                    published_start_at: '2023-11-30 08:00:00'
                    published_end_at: '2024-12-01 18:00:00'
                    updated_at: '2024-02-07 00:12:26'
                    created_at: '2024-02-05 17:55:34'
                success: true
                code: 200
                pagination: null
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '404':
          $ref: '#/components/responses/404CommentAndPostResourceNotFoundResponse'
  '/posts/{postId}/comments/{id}/unpublish':
    post:
      summary: Unpublishes a comment
      description: >
        - Unpublishes a comment associated with a post.

        - Only posters have permission to unpublish a comment

        - Nickname and avatar information of the commenter are fetched from the
        CRM API and are not managed by the Social API.
      tags:
        - Comment-related endpoints
      operationId: unpublish-a-comment
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: postId
          in: path
          description: The ID of the post associated with the comment.
          required: true
          schema:
            type: string
            example: 01hnw7nqm19g9302c7wq391nvm
        - name: id
          in: path
          description: The ID of the comment to publish.
          required: true
          schema:
            type: string
            example: 01hnwcp6kd9a50w0hy2q0tg0zd
      responses:
        '200':
          description: Comment unpublished successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResource'
              example:
                data:
                  id: 01hnwcp6kd9a50w0hy2q0tg0zd
                  user_id: 16614
                  nickname: hoctran
                  avatar: >-
                    https://stage.user.com/file/fileconsumer/doo/selector/downloader/phase/download_file/type/images/file_id/phpv2w4cF
                  comment: Your comment goes here
                  is_hidden: false
                  updated_at: '2024-02-07 19:08:15'
                  created_at: '2024-02-05 19:23:13'
                  post:
                    id: 01hnw7nqm19g9302c7wq391nvm
                    admin_uuid: 9295de24-404b-408f-8ec5-1a5f3224a244
                    type: premium
                    content: Test content
                    is_published: true
                    published_start_at: '2023-11-30 08:00:00'
                    published_end_at: '2024-12-01 18:00:00'
                    updated_at: '2024-02-07 00:12:26'
                    created_at: '2024-02-05 17:55:34'
                success: true
                code: 200
                pagination: null
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '404':
          $ref: '#/components/responses/404CommentAndPostResourceNotFoundResponse'
...
components:
  parameters:
...
    X-Social-Api-Key:
      name: X-Social-Api-Key
      in: header
      description: |
        An unique string provided by Tranvelocity when you create an account. You
        must include this API key in the specified HTTP request header for every
        API request to authenticate your requests.
      required: true
      schema:
        type: string
        example: FSFSDFSDFDSpHYv7kMeb0yVkn2343242L091nI5GLkKPfAt234324bmNEaR
    X-Social-Timestamp:
      name: X-Social-Timestamp
      in: header
      description: |
        An Unix epoch value representing the current time. You must include this
        timestamp in the specified HTTP request header for every API request to
        prevent replay attacks and ensure request freshness.
      required: true
      schema:
        type: integer
        example: 1595907741
    X-Social-Signature:
      name: X-Social-Signature
      in: header
      description: >
        A signature that is HMAC-encoded with SHA-512 digest. You must include
        this

        signature in the specified HTTP request header for every API request to

        ensure data integrity and authenticity.
      required: true
      schema:
        type: string
        example: 5d714ce2385d1cf96260806a50f3f7e68b928cc8
...